рефакторинг модуля утилит
