+++ src/main.js
@@ -1,13 +1,10 @@
 import {IterationCount} from "./constants.js";
-import {render} from "./util.js";
+import {render} from "./Utils/render.js";
 import BoardPresenter from "./presenter/board.js";
 import SiteMenuView from "./View/site-menu.js";
 import FilterView from "./View/filter.js";
-
-
 import {generateTask} from "./mock/task.js";
 import {generateFilter} from "./mock/filter.js";
-
 
 export const tasks = new Array(IterationCount.CARD).fill().map(generateTask);
 const filters = generateFilter(tasks);
@@ -21,3 +18,4 @@ render({container: siteMainElement, child: new FilterView(filters)});
 const presenter = new BoardPresenter(siteMainElement);
 
 presenter.init(tasks);
+
